samples/agent-catalog/3p-agent-samples/foundry-agent-service-sdk/saifr-comm-compliance-agent/python/requirements.txt
jsonref==1.1.0
azure-core==1.34.0
azure-ai-projects==1.0.0b10
azure-identity==1.23.0